requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0

